bet done test

+++ contracts/BetHolderContract.sol
@@ -3,7 +3,7 @@ pragma solidity 0.5.12;
 import "@openzeppelin/contracts/ownership/Ownable.sol";
 
 contract BetHolderContract is Ownable {
-    mapping(address => uint) public players;
+    mapping(address => uint256) public players;
     constructor() public {
 
     }
